Jump to Definition Crash
Bug was reported that RepDev Crash when Ctrl-Shift-G is pressed.  I was
finally able to duplicate the bug, 11 months later.  Of course it would
have helped if I'd read the notes clearly :/
The issue was caused by the code not checking if the tab is an Editor
Composit or Report Composit.  Not sure if the best way to check for it is
if(tf.getControl().toString().equals("EditorComposite {}")).  Guess I can
always change it if I find the correct way.