Quick compare local file to production file
Not sure why I explicitly not allowed the local files to be compared to
the production file with the right click "Compare to Production" feature.
If we can select a local file and a production file and select "Compare
Two Files", then I do not see any issues with this.