Making my life easier
Since most everyone are using 64 bits now, I am defaulting the repo to
x64.  Also adding a batch file to help with troubleshooting ssh issues.
Updating .gitignore as well as .classpath and moving everything needed
into the repo so that I can click build and everything will be ready
to go, and all I have to do is zip it up for download.  This way I do
not need move required files into the RepDev folder.  Hope this does
not mess anyone up.