RepDev Single Sign-On
If you enable this feature, it will remember the last credential used
for each of the SYMs that you have configured.  If your AIX password
changed, you can mass update all of the SYMs that uses that server.
I call this RepDev SSO, not to be confused with Quest SSO.

If you currently run your own build and enable the DEVELOPER mode
to save passwords, the password is saved in plain text and is not
secure.  This new sso feature will require you to set a RepDev
Password.  This password is used to encrypt (AES256) your AIX and
Teller Passwords at rest.  RepDev will ask you for the password
each time you launch it, hence single sign on.

To enable this feature, click on Tools, Options, under the Server
Options tab, check the box for “Use Single Sign-On”.  It will
prompt you to create a RepDev password.  There is a minimum
requirement for the password complexity.  If your AIX password
expired, change it on the Symitar Server first.  To update your
saved AIX Password in RepDev, click on Tools, Change AIX Password.
It will prompt you for the server name, then tell you how many and
what SYMs are using that server.  Enter your new AIX Password and
you’re good to go.  If you forget your RepDev password, disable
sso and re-enable it to start fresh.