Run report Shell - fixed size
The Run Report Shell auto expands as prompts are answered. Typically
this does not pose any problems, but some RepGens may have a lot of
prompts or multi lined prompts that will scroll off the screen and
make it impossible to answer all the prompts.  This patch will make
the window fixed size and allow the scrollbars to do it's job.