Snippet Manager - Sorts alphabetically
If you create or add snippets to the snippets directory, it
may be out of order in RepDev.  For some reason, RepDev uses
the Create Date to pull them in.  This will sort the List of
snippets so that it does not matter when the snippets are
created or saved.