Date Archive
If you try to date archive multiple files with the first 24
characters of the file names being the same, the first file
will succeed, but the subsequent files will fail due the
file name ending up being the same.  Updated the algorithm
to prepend 1-30 in front of the .mmddyy if the new file name
already exists.