Password will expire
My RepDev will enter the Symulate mode if my password is due to
expire, that is why this was programmed to give an error and
terminate gracefully.  I noticed, due to various parameters, it
may log you in sucessfully.  Due to that, I updated it to
continue on and stop only if goes into Symulate Mode.  I have
experienced RepDev still freeze.  I guess this is a trade off for
users that want to be able to continue without having to force
a password change.