Double clicking on CALL crashes if procedure no longer exist
Zebulan reported that, if an include file was modified or deleted, and you double
click on a CALL to a procedure that is no longer there, RepDev will
crash.  See Issue #43.  This has been patched to Ignore the exception,
but it will still try to open the Include File and probably go to the
bottom of the file.