Read Only File System
Chris Thompson reported that his Symitar box had the wrong
File permissions and did not have write access.  He was saving
files not knowing that the files were not actually saved
because RepDev did not give him any errors.  This patch will
detect "Unable to save on Output Open" and popup an error
message "File not saved.  Read Only !!" to alert the user.