Enhanced Password due to expire
Previously I patched RepDev to allow RepDev to continue to
log in if the user's AIX PAssword is due to expire.  For
some users, RepDev cannot log in and ends up in Symulate
mode.  At that point RepDev displays the message
"WINDOWSLEVEL not set to 3." which is not very intuitive.
I Updated the message to say "RepDev is in Symulate mode
because your password is due to expire.  Please change
it now.".